Move :core Configuration Cache extensions to internal package
In separate module.
