+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/services/Environment.kt
@@ -17,7 +17,7 @@
 package org.gradle.configurationcache.services
 
 import org.gradle.configurationcache.extensions.filterKeysByPrefix
-import org.gradle.configurationcache.extensions.getBroadcaster
+import org.gradle.internal.extensions.core.getBroadcaster
 import org.gradle.configurationcache.extensions.uncheckedCast
 import org.gradle.initialization.Environment
 import org.gradle.internal.event.ListenerManager
