+++ platforms/core-configuration/configuration-problems-base/src/main/kotlin/org/gradle/internal/configuration/problems/DocumentRegistryExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.configurationcache.extensions
+package org.gradle.internal.configuration.problems
 
-import org.gradle.groovy.scripts.ScriptSource
-import java.net.URI
+import org.gradle.api.internal.DocumentationRegistry
 
 
-/**
- * Same as `resource.location.uri`.
- */
-internal
-val ScriptSource.uri: URI?
-    get() = resource.location.uri
+fun DocumentationRegistry.documentationLinkFor(section: DocumentationSection) =
+    configurationCacheLinkFor(section.anchor)
+
+
+fun DocumentationRegistry.configurationCacheLinkFor(anchor: String): String =
+    getDocumentationFor("configuration_cache", anchor)
