+++ settings.gradle.kts
@@ -113,6 +113,7 @@ val core = platform("core") {
         subproject("base-services-groovy")
         subproject("configuration-cache")
         subproject("configuration-problems-base")
+        subproject("core-extensions-kt")
         subproject("declarative-dsl-api")
         subproject("declarative-dsl-core")
         subproject("declarative-dsl-provider")
