+++ testing/architecture-test/src/test/java/org/gradle/architecture/test/ArchUnitFixture.java
@@ -76,6 +76,7 @@ public interface ArchUnitFixture {
     DescribedPredicate<JavaClass> classes_not_written_in_kotlin = resideOutsideOfPackages(
         "org.gradle.configurationcache..",
         "org.gradle.internal.configuration.problems..",
+        "org.gradle.internal.extensions.core..",
         "org.gradle.internal.serialize.graph..",
         "org.gradle.kotlin..",
         "org.gradle.internal.declarativedsl..",
